pyproject.toml
[project]
name = "closeclaw"
version = "0.1.0"
description = "AI Agent on Telegram, powered by Kimi"
readme = "README.md"
authors = [
    { name = "Justin Wong", email = "justin.w.xd@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "kosong>=0.46.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.3",
    "rich>=13.0.0",
    "click>=8.0.0",
    "python-telegram-bot>=21.0",
    "kimi-cli>=1.27",
    "pykaos>=0.7.0",
]

[project.scripts]
closeclaw = "closeclaw.cli.main:cli"

[dependency-groups]
dev = [
    "prek>=0.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
